# src/app/openapi_spec.json
{"openapi": "3.0.2", "info": {"title": "FastAPI", "description": "\n# Before consuming the endpoints set the variable `MODEL_DIRECTORY` in `.env` to a directory that contains model checkpoints (.ckpt). Models outside of this directory can't be used for inference via endpoints.\n\nCommand for running the server:\n### `(venv) username@pc:~/lumen-geoguesser$ python3 src/app/main.py`\n\n", "version": "0.1.0"}, "paths": {"/models": {"get": {"tags": ["available models"], "summary": "Get Models", "description": "Returns names of all available models on the server. You must use model names for all `POST` request predictions. Model name is a stem of the model checkpoint, e.g. model with filename `my_model.ckpt` has the name `my_model`. Models are fetched from the directory `MODEL_DIRECTORY` which is defined in the `.env` file. Only models with the extension `MODEL_EXTENSION` are fetched. Model names will be returned instead of the model filenames.\n", "operationId": "get_models_models_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get Models Models Get", "type": "array", "items": {"type": "string"}}}}}}}}, "/model/{model_name}/predict-images": {"post": {"tags": ["predict"], "summary": "Predict Images", "description": "Infers latitude and longitude for multiple images. If you have a group of images where each image represents one cardinal direction (north, east, south, and west) (\"0.jpg\", ... , \"270.jpg\") you should use the `/model/{model_name}/predict-cardinal-images endpoint\n", "operationId": "predict_images_model__model_name__predict_images_post", "parameters": [{"required": true, "schema": {"title": "Model Name", "type": "string"}, "name": "model_name", "in": "path"}], "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_predict_images_model__model_name__predict_images_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Predict Images Model  Model Name  Predict Images Post", "type": "array", "items": {"$ref": "#/components/schemas/PredictImagesResponse"}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/model/{model_name}/predict-cardinal-images": {"post": {"tags": ["predict"], "summary": "Predict Cardinal Images", "description": "Infers latitude and longitude for a single location which is defined by exactly 4 images, each for one cardinal direction. Exactly 4 images must be sent and each image filename must match image's cardinal direction (\"0.jpg\", \"90.jpg\", \"180.jpg\", \"270.jpg\"). E.g northen image should be named \"0.jpg\". This structure is the same as the structure of the original dataset.\n", "operationId": "predict_cardinal_images_model__model_name__predict_cardinal_images_post", "parameters": [{"required": true, "schema": {"title": "Model Name", "type": "string"}, "name": "model_name", "in": "path"}], "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_predict_cardinal_images_model__model_name__predict_cardinal_images_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Predict Cardinal Images Model  Model Name  Predict Cardinal Images Post", "type": "array", "items": {"$ref": "#/components/schemas/PredictImagesCardinalResponse"}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/model/{model_name}/predict-directory": {"post": {"tags": ["predict"], "summary": "Predict Dataset", "description": "Infers latitude and longitude for all images in the directory (`dataset_directory_path`) which contains subdirectories (uuid) with images for each cardinal direction. This structure is the same as the structure of the original dataset. Exactly 4 images must be sent per subdirectory and each image filename must match image's cardinal direction (\"0.jpg\", \"90.jpg\", \"180.jpg\", \"270.jpg\"). E.g northen image should be named \"0.jpg\". If `csv_filename` is provided, results will also be saved to a .csv file.\n", "operationId": "predict_dataset_model__model_name__predict_directory_post", "parameters": [{"required": true, "schema": {"title": "Model Name", "type": "string"}, "name": "model_name", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/PostPredictDatasetRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Predict Dataset Model  Model Name  Predict Directory Post", "type": "array", "items": {"$ref": "#/components/schemas/PredictDirectoryReponse"}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"Body_predict_cardinal_images_model__model_name__predict_cardinal_images_post": {"title": "Body_predict_cardinal_images_model__model_name__predict_cardinal_images_post", "required": ["images"], "type": "object", "properties": {"images": {"title": "Images", "type": "array", "items": {"type": "string", "format": "binary"}}}}, "Body_predict_images_model__model_name__predict_images_post": {"title": "Body_predict_images_model__model_name__predict_images_post", "required": ["images"], "type": "object", "properties": {"images": {"title": "Images", "type": "array", "items": {"type": "string", "format": "binary"}}}}, "HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "PostPredictDatasetRequest": {"title": "PostPredictDatasetRequest", "required": ["dataset_directory_path"], "type": "object", "properties": {"dataset_directory_path": {"title": "Dataset Directory Path", "type": "string"}, "csv_filename": {"title": "Csv Filename", "type": "string"}}}, "PredictDirectoryReponse": {"title": "PredictDirectoryReponse", "required": ["uuid", "latitude", "longitude"], "type": "object", "properties": {"uuid": {"title": "Uuid", "type": "string", "format": "uuid"}, "latitude": {"title": "Latitude", "type": "number"}, "longitude": {"title": "Longitude", "type": "number"}}}, "PredictImagesCardinalResponse": {"title": "PredictImagesCardinalResponse", "required": ["latitude", "longitude"], "type": "object", "properties": {"latitude": {"title": "Latitude", "type": "number"}, "longitude": {"title": "Longitude", "type": "number"}}}, "PredictImagesResponse": {"title": "PredictImagesResponse", "required": ["latitude", "longitude"], "type": "object", "properties": {"latitude": {"title": "Latitude", "type": "number"}, "longitude": {"title": "Longitude", "type": "number"}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}, "tags": [{"name": "available models", "description": "All available models from the `MODEL_DIRECTORY` directory defined in the `.env` file"}, {"name": "predict", "description": "Endpoints used for predicting latitude and longitudes for given data.\nCurl example for multiple images:\n\n```\ncurl -i \\\n-F \"images=@data/raw/images/train/e788b3d1-9d20-466c-9dee-97982f0f9a3b/0.jpg\" \\\n-F \"images=@data/raw/images/train/e788b3d1-9d20-466c-9dee-97982f0f9a3b/0.jpg\" \\\nhttp://0.0.0.0:8090/model/Golf_76__haversine_0.0098__val_acc_0.47__val_loss_1.98__05-04-03-36-32/predict-images\n```\n"}]}